Typescript init
