+++ .prettierrc
@@ -1,6 +1,6 @@
 {
   "semi": false,
-  "singleQuote": true,
+  "singleQuote": false,
   "trailingComma": "es5",
   "printWidth": 140
 }
