--- src/components/Logo.js
@@ -1,12 +1,0 @@
-import React from 'react'
-import styled from 'styled-components'
-
-const PageTitle = styled.h1`
-  font-size: 24px;
-  margin: 0;
-  font-weight: 500;
-`
-
-const Logo = () => <PageTitle>Theo Dammaretz</PageTitle>
-
-export default Logo
