+++ src/types/MediumPost.d.ts
@@ -1,0 +1,11 @@
+export interface MediumPost {
+  title: string
+  uniqueSlug: string
+  firstPublishedAt: string
+  author: { username: string }
+  virtuals: MediumVirtuals
+}
+
+export interface MediumVirtuals {
+  tags: { name: string }[]
+}
