+++ src/types/StaticQuery.d.ts
@@ -1,0 +1,7 @@
+import { SiteMeta } from "./siteMeta"
+import { MediumPost } from "./MediumPost"
+
+export interface StaticQueryResult {
+  site: { siteMetadata: SiteMeta }
+  allMediumPost: { edges: { node: MediumPost }[] }
+}
