+++ src/types/typings.d.ts
@@ -1,0 +1,4 @@
+declare module '*.png';
+declare module '*.jpg';
+declare module '*.json';
+declare module '*.svg';
