+++ src/utils/styled-components.ts
@@ -1,0 +1,8 @@
+import * as styledComponents from "styled-components"
+
+import { Theme } from "./theme"
+
+const { default: styled, css, injectGlobal, keyframes, ThemeProvider, withTheme } = styledComponents as styledComponents.ThemedStyledComponentsModule<Theme>
+
+export { css, injectGlobal, keyframes, ThemeProvider, withTheme }
+export default styled
