--- src/utils/theme.js
@@ -1,7 +1,0 @@
-export default {
-    primary: "#EF476F",
-    secondary: "#05B183",
-    tertiary: "#FFD166",
-    darkText: "#053342",
-    lightText: "#365F6D",
-}
